requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
paddleocr==2.8.1
opencv-python==4.10.0.84
python-docx==1.1.2
numpy==1.26.4
jinja2==3.1.4
